switch with enums
In Swift, enumerations (enums) are a powerful feature that allows you to define a common type for a group of related values. When you have an enum, you often need to perform different actions based on the specific case of the enum. This is where the switch statement comes in handy.