switch-statement
Another type of conditional statement that exists in Swift is the switch statement. The switch statement is a popular programming tool used to check the value of a given expression against multiple cases. The switch statement is a lot more powerful in Swift than it is in other programming languages, thus we’ll be dedicating the next few exercises to explore its features.

Unlike the if statement, a switch statement does not check for the value of a condition and instead finds and matches a case to a given expression.