Create switch-statement-interval-matching
One super power that the switch statement possesses, is its ability to match values to an expression that exist within intervals. An interval denotes a range used for checking whether a given value lies within that range.

The closed range operator signifies an inclusive range where the first and last values are included in the sequence.